+++ mountpoint-s3-fs/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+* Update to latest S3 client. ([#1793](https://github.com/awslabs/mountpoint-s3/pull/1793))
+
 ## v0.9.1 (March 9, 2026)
 
 * Fix a race condition where concurrent operations after closing a truncated file could result in I/O errors on subsequent reads. The issue was introduced in `mountpoint-s3-fs` v0.9.0. ([#1781](https://github.com/awslabs/mountpoint-s3/pull/1781))